djangoProject_VMS/vendor_management/requirement.txt
asgiref==3.7.2
deap==1.4.1
Django==4.2.8
djangorestframework==3.14.0
et-xmlfile==1.1.0
pytz==2023.3.post1
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
xlrd==2.0.1
xlwt-future==0.8.0
